pyproject.toml
[project]
name = "p52"
version = "0.0.1"
description = "Automatically launch p5js sketches written in python"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pywebview",
    "transcrypt",
    "websockets",
    "watchdog"
]

[tool.setuptools.package-data]
"p52.static" = ["index.html"]